+++ selendroid-grid-plugin/src/main/java/io/selendroid/grid/SelendroidNodeHtmlRenderer.java
@@ -27,9 +27,6 @@ import org.openqa.selenium.remote.DesiredCapabilities;
  */
 public class SelendroidNodeHtmlRenderer implements HtmlRenderer {
   public static final String BROWSER_NAME = "selendroid";
-  public static final String LOCALE = "locale";
-  public static final String SDK_VERSION = "sdkVersion";
-  public static final String AUT = "aut";
 
   private SelendroidSessionProxy proxy;